siem_node/requirements.txt
requests
psutil
pyudev
inotify-simple
pycryptodome
pyyaml
python-socketio
scapy
dnspython